backend/pyproject.toml
[project]
name = "assistente-social-media-backend"
version = "0.1.0"
description = "Backend FastAPI para geração de marketing com agentes IA"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "anthropic>=0.40",
    "google-genai>=0.8",
    "supabase>=2.10",
    "asyncpg>=0.29",
    "sse-starlette>=2.1",
    "python-jose[cryptography]>=3.3",
    "pillow>=10.4",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
